@@ -1,10 +1,10 @@
 name: pspdfkit_example
 description: Demonstrates how to use the pspdfkit plugin.
-version: 3.3.0
+version: 3.5.0
 homepage: https://pspdfkit.com/
 environment:
   sdk: '>=2.12.0 <3.0.0'
-  flutter: '>=3.0.4'
+  flutter: '>=3.3.9'
 
 dependencies:
   flutter:
@@ -13,6 +13,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
+  http: ^0.13.5
+  qr_code_scanner: ^1.0.1
+  url_launcher: ^6.1.7
+
 
 dev_dependencies:
   flutter_lints: ^2.0.1
