@@ -1,6 +1,6 @@
 name: pspdfkit_flutter
 description: A Flutter plugin providing a feature-rich PDF viewing and editing experience to your users with the powerful PSPDFKit PDF SDK.
-version: 3.3.0
+version: 3.5.0
 homepage: https://pspdfkit.com/
 repository: https://github.com/PSPDFKit/pspdfkit-flutter
 issue_tracker: https://support.pspdfkit.com/hc/en-us/requests/new
@@ -9,11 +9,13 @@ environment:
   sdk: '>=2.12.0 <3.0.0'
   # Flutter versions prior to 1.12 did not support the
   # flutter.plugin.platforms map.
-  flutter: '>=3.0.4'
+  flutter: '>=3.3.9'
 
 dependencies:
+  build_runner: ^2.2.1
   flutter:
     sdk: flutter
+  mockito: ^5.3.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
