@@ -8,6 +8,9 @@
  */
 package com.pspdfkit.flutter.pspdfkit;
 
+import static com.pspdfkit.flutter.pspdfkit.util.Preconditions.requireNotNullNotEmpty;
+import static io.flutter.util.Preconditions.checkNotNull;
+
 import android.content.Context;
 import android.util.Log;
 
@@ -15,6 +18,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.StyleRes;
 
+import com.pspdfkit.configuration.PdfConfiguration;
 import com.pspdfkit.configuration.activity.PdfActivityConfiguration;
 import com.pspdfkit.configuration.activity.ThumbnailBarMode;
 import com.pspdfkit.configuration.activity.UserInterfaceViewMode;
@@ -31,9 +35,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.EnumSet;
 import java.util.HashMap;
-
-import static com.pspdfkit.flutter.pspdfkit.util.Preconditions.requireNotNullNotEmpty;
-import static io.flutter.util.Preconditions.checkNotNull;
+import java.util.List;
 
 class ConfigurationAdapter {
     private static final String LOG_TAG = "ConfigurationAdapter";
@@ -119,12 +121,12 @@ class ConfigurationAdapter {
      */
     @Deprecated
     private static final String SHOW_DOCUMENT_LABEL = "showDocumentLabel";
-    /** 
+    /**
      * @deprecated This key word was deprecated with PSPDFKit for Fluttter 3.1.
      * Use {@code FIRST_PAGE_ALWAYS_SINGLE} instead, which replaces it.
      */
     private static final String IS_FIRST_PAGE_ALWAYS_SINGLE = "isFirstPageAlwaysSingle";
-    /** 
+    /**
      * @deprecated This key word was deprecated with PSPDFKit for Fluttter 3.1.
      * Use {@code SHOW_BOOKMARKS_ACTION} instead, which replaces it.
      */
@@ -136,7 +138,7 @@ class ConfigurationAdapter {
     private static final String PAGE_TRANSITION_SCROLL_PER_SPREAD = "scrollPerSpread";
     private static final String PAGE_TRANSITION_SCROLL_CONTINUOUS = "scrollContinuous";
     private static final String PAGE_TRANSITION_CURL = "curl";
-    
+
     // Document Presentation Values
     private static final String PAGE_MODE_AUTOMATIC = "automatic";
     private static final String PAGE_MODE_SINGLE = "single";
@@ -177,8 +179,14 @@ class ConfigurationAdapter {
     private static final String SHOW_THUMBNAIL_BAR_SCRUBBER_BAR = "scrubberBar";
     private static final String SHOW_THUMBNAIL_BAR_SCROLLABLE = "scrollable";
 
-    @NonNull private final PdfActivityConfiguration.Builder configuration;
-    @Nullable private String password = null;
+    // Instant Options
+    private static final String ENABLE_INSTANT_COMMENTS = "enableInstantComments";
+
+    @NonNull
+    private final PdfActivityConfiguration.Builder configuration;
+    @Nullable
+    private String password = null;
+    private boolean enableInstantComments = false;
 
     ConfigurationAdapter(@NonNull Context context,
                          @Nullable HashMap<String, Object> configurationMap) {
@@ -352,6 +360,10 @@ class ConfigurationAdapter {
                 configureAutosaveEnabled(!(Boolean) configurationMap.get(key));
             }
             configurationAnnotationTools();
+            key = getKeyOfType(configurationMap, ENABLE_INSTANT_COMMENTS, Boolean.class);
+            if (key != null) {
+                enableInstantComments = (boolean) configurationMap.get(key);
+            }
         }
     }
 
@@ -679,27 +691,27 @@ class ConfigurationAdapter {
                 case SETTINGS_MENU_ITEM_THEME:
                 case SETTINGS_MENU_ITEM_ANDROID_THEME:
                     settingsMenuItemTypes.add(SettingsMenuItemType.THEME);
-                break;
+                    break;
                 case SETTINGS_MENU_ITEM_SCREEN_AWAKE:
                 case SETTINGS_MENU_ITEM_ANDROID_SCREEN_AWAKE:
                     settingsMenuItemTypes.add(SettingsMenuItemType.SCREEN_AWAKE);
-                break;
+                    break;
                 case SETTINGS_MENU_ITEM_PAGE_LAYOUT:
                 case SETTINGS_MENU_ITEM_ANDROID_PAGE_LAYOUT:
                     settingsMenuItemTypes.add(SettingsMenuItemType.PAGE_LAYOUT);
-                break;
+                    break;
                 case SETTINGS_MENU_ITEM_PAGE_TRANSITION:
                     settingsMenuItemTypes.add(SettingsMenuItemType.PAGE_TRANSITION);
-                break;
+                    break;
                 case SETTINGS_MENU_ITEM_SCROLL_DIRECTION:
                     settingsMenuItemTypes.add(SettingsMenuItemType.SCROLL_DIRECTION);
-                break;
+                    break;
                 case SETTINGS_MENU_ITEM_IOS_APPEARANCE:
                 case SETTINGS_MENU_ITEM_IOS_BRIGHTNESS:
                 case SETTINGS_MENU_ITEM_IOS_PAGE_MODE:
                 case SETTINGS_MENU_ITEM_IOS_SPREAD_FITTING:
                     // NO-OP. Only supported on iOS.
-                break;
+                    break;
                 default:
                     throw new IllegalArgumentException("Undefined settings menu item " + menuType);
             }
@@ -736,13 +748,13 @@ class ConfigurationAdapter {
      * When reading configuration options, we check not only for the given configuration string,
      * but also for a string with the `android` prefix. For instance if the user enters
      * `androidPageScrollDirection`, it is considered a valid string equal to `pageScrollDirection`.
-     * 
+     * <p>
      * When documenting, we always prefer configuration option strings:
-     * 
+     * <p>
      * - No prefix          : If the key works for both iOS and Android.
      * - `android` prefix   : If the key works only for Android.
      * - `iOS` prefix       : If the key works only for iOS.
-     */ 
+     */
     private String addAndroidPrefix(String key) {
         // Capitalize the first letter.
         String cap = String.valueOf(key.charAt(0)).toUpperCase() + key.substring(1);
@@ -751,8 +763,8 @@ class ConfigurationAdapter {
 
     @Nullable
     private <T> String getKeyOfType(@NonNull HashMap<String, Object> configurationMap,
-                                @NonNull String key,
-                                @NonNull Class<T> clazz) {
+                                    @NonNull String key,
+                                    @NonNull Class<T> clazz) {
         if (containsKeyOfType(configurationMap, key, clazz)) {
             return key;
         }
@@ -802,6 +814,19 @@ class ConfigurationAdapter {
     }
 
     PdfActivityConfiguration build() {
+//      Turn on Instant comments;
+        if (enableInstantComments) {
+            PdfConfiguration pdfConfiguration = this.configuration.build().getConfiguration();
+
+            final List<AnnotationTool> annotationTools = pdfConfiguration
+                    .getEnabledAnnotationTools();
+            // Explicitly enable Instant comment markers if your server supports this feature.
+            annotationTools.add(AnnotationTool.INSTANT_COMMENT_MARKER);
+            annotationTools.add(AnnotationTool.INSTANT_HIGHLIGHT_COMMENT);
+            this.configuration
+                    .enabledAnnotationTools(annotationTools);
+        }
+
         return configuration.build();
     }
 }
