@@ -2,6 +2,6 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   test('Placeholder test', () async {
-    expect('1', '1');
+    expect('0', '0');
   });
 }
